// entsnap/src/app/models/gpsGeocoder.ts
export interface GpsGeocoder {
    latitude?: number | any;
    longitude?: number | any;
    accuracy?: number;
    heading?: any;
    street_address?: any;
    route?: string;
    locality?: string;
    country?: string;
    administrative_area_level_1?: string;
    administrative_area_level_2?: string;
    administrative_area_level_3?: string;
}